LayoutTests/media/video-controls-show-on-kb-or-ax-event-expected.txt

This tests:
1. Initially, toolbar is visible, "Show Controls" button is not.
2. After the video controls fade out, they can be shown when VoiceOver or a keyboard user clicks the invisible "Show Controls" button.
3. When the video is paused, the controls become visible and the "Show Controls" button is hidden.

PASS
PASS
PASS
PASS
PASS
PASS
PASS
PASS
